updated/improved script
Hey Sean,

I Changed script from if statements to case and cleaned up the script.

Struggling to understand how I can use a loop here for the case. When i enter an option I want to be able to go back to the main menu instead of quitting.

Thanks dude,

salloweezy